Add voting power to validators (#2)
To enable delegation we need every validator to have a specific voting power based on its staked and delegated balance.

A voting Power struct was added
A voting power cache was added to the contract store
All consensus methods that worked with a number of validators are now modified to work with the voting power
Tests were updated for all changed functionalities and new tests were added for the new functionalities

NB! Fetcher currently mocks the same staked balance for all validators. This must be changed whenever the staking contracts are implemented.